# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TikTok Downloader API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  phone: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserResponse {
  id: number;
  name: string;
  email: string;
  phone: string;
  createdAt: string;
  downloadsCount: number;
  hasActiveSubscription: boolean;
}

export interface AuthResponse {
  user: UserResponse;
  token: string;
}

export interface SubscribeBody {
  /** M-Pesa phone number to receive the STK push (e.g. 07XXXXXXXX or 2547XXXXXXXX). Defaults to the user's registered phone. */
  phone?: string;
}

export interface DownloadBody {
  url: string;
}

export interface DownloadResponse {
  downloadUrl: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  remainingFreeDownloads: number;
}

export interface PaymentRequiredResponse {
  error: string;
  subscriptionPrice: number;
  currency: string;
}

export interface DownloadRecord {
  id: number;
  url: string;
  downloadedAt: string;
}

export interface SubscriptionStatus {
  isActive: boolean;
  /** @nullable */
  expiresAt?: string | null;
  downloadsCount: number;
  remainingFreeDownloads: number;
  subscriptionPrice: number;
  currency: string;
}

export interface PaymentInitResponse {
  /** True when the M-Pesa STK push was successfully sent to the user's phone. */
  stkSent: boolean;
  /** Human-readable status message to show the user. */
  message: string;
  amount: number;
  currency: string;
}

export interface PaylorCallbackBody {
  reference: string;
  status: string;
  amount: number;
  /** @nullable */
  phone?: string | null;
}

export interface AdminUserRecord {
  id: number;
  name: string;
  email: string;
  phone: string;
  createdAt: string;
  downloadsCount: number;
  isSuspended: boolean;
  /** @nullable */
  subscriptionStatus?: string | null;
  /** @nullable */
  subscriptionExpiresAt?: string | null;
}

export interface AdminSettings {
  subscriptionPrice: number;
  currency: string;
  paylorApiKey: string;
  paylorApiUrl: string;
  paylorChannelId: string;
  adminKey: string;
  freeDownloadsPerUser: number;
}

export interface AdminSettingsUpdate {
  /** @nullable */
  subscriptionPrice?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  paylorApiKey?: string | null;
  /** @nullable */
  paylorApiUrl?: string | null;
  /** @nullable */
  paylorChannelId?: string | null;
  /** @nullable */
  adminKey?: string | null;
  /** @nullable */
  freeDownloadsPerUser?: number | null;
}

export interface AdminStats {
  totalUsers: number;
  activeSubscribers: number;
  totalDownloads: number;
  revenueThisMonth: number;
  newUsersThisMonth: number;
}