# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TikTok Downloader API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: List all users
      description: Returns all registered users with their subscription status. Requires admin key.
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUserRecord"
        "401":
          description: Invalid or missing admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get admin settings
      description: Returns configurable admin settings (subscription price, Paylor keys). Requires admin key.
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
        "401":
          description: Invalid or missing admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update admin settings
      description: Update configurable admin settings. Requires admin key.
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
        "401":
          description: Invalid or missing admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/upgrade:
    post:
      operationId: adminUpgradeUser
      tags: [admin]
      summary: Upgrade user to Pro
      description: Activates a 1-month Pro subscription for the user. Requires admin key.
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User upgraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Invalid or missing admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/suspend:
    post:
      operationId: adminSuspendUser
      tags: [admin]
      summary: Suspend a user
      description: Suspends the user account, revoking all active tokens. Requires admin key.
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Invalid or missing admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/unsuspend:
    post:
      operationId: adminUnsuspendUser
      tags: [admin]
      summary: Unsuspend a user
      description: Lifts a suspension from the user account. Requires admin key.
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unsuspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Invalid or missing admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user
      description: Permanently deletes the user and all their data. Requires admin key.
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Invalid or missing admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get admin dashboard stats
      description: Returns overview stats for the admin dashboard. Requires admin key.
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Invalid or missing admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    AdminUserRecord:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
        downloadsCount:
          type: integer
        isSuspended:
          type: boolean
        subscriptionStatus:
          type: ["string", "null"]
        subscriptionExpiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - createdAt
        - downloadsCount
        - isSuspended
    AdminSettings:
      type: object
      properties:
        subscriptionPrice:
          type: number
        currency:
          type: string
        paylorApiKey:
          type: string
        paylorApiUrl:
          type: string
        paylorChannelId:
          type: string
        adminKey:
          type: string
        freeDownloadsPerUser:
          type: integer
      required:
        - subscriptionPrice
        - currency
        - paylorApiKey
        - paylorApiUrl
        - paylorChannelId
        - adminKey
        - freeDownloadsPerUser
    AdminSettingsUpdate:
      type: object
      properties:
        subscriptionPrice:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        paylorApiKey:
          type: ["string", "null"]
        paylorApiUrl:
          type: ["string", "null"]
        paylorChannelId:
          type: ["string", "null"]
        adminKey:
          type: ["string", "null"]
        freeDownloadsPerUser:
          type: ["integer", "null"]
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeSubscribers:
          type: integer
        totalDownloads:
          type: integer
        revenueThisMonth:
          type: number
        newUsersThisMonth:
          type: integer
      required:
        - totalUsers
        - activeSubscribers
        - totalDownloads
        - revenueThisMonth
        - newUsersThisMonth